Alterado estrutura de cell

+++ Cell.h
@@ -6,7 +6,7 @@ typedef struct cell Cell;
 /* Getters */
 int cell_getI(Cell *cell); /* Retorna a posicao I (linha) da celula */
 int cell_getJ(Cell *cell); /* Retorna a posicao J (coluna) da celula */
-int cell_getK(Cell *cell); /* Retorna o valor K contido na celular */
+float cell_getK(Cell *cell); /* Retorna o valor K contido na celular */
 Cell *cell_getNextOfRow(Cell *cell); /* Retorna o ponteiro para a proxima celula da linha */
 Cell *cell_getNextOfCol(Cell *cell); /* Retorna o ponteiro para a proxima celula da coluna */
 /* end of Getters */
@@ -14,7 +14,7 @@ Cell *cell_getNextOfCol(Cell *cell); /* Retorna o ponteiro para a proxima celula
 /* Setters */
 void cell_setI(Cell *cell, int i); /* Define a posicao I da celula */
 void cell_setJ(Cell *cell, int j); /* Define a posicao J da celula */
-void cell_setK(Cell *cell, int k); /* Define o valor K da celula */
+void cell_setK(Cell *cell, float k); /* Define o valor K da celula */
 void cell_setNextOfRow(Cell *cell, Cell *ncell); /* Define o proxima celula da linha */
 void cell_setNextOfCol(Cell *cell, Cell *ncell); /* Define a proxima celula da coluna */
 /* end of Setters */
